(backup)

+++ seed_ok.sh
@@ -4,14 +4,15 @@ if [ -z "$1" ]; then
 	exit 1
 fi
 BIN=./push_swap
-C_BIN=./checker
-if [ ! -f "$C_BIN" ]; then
-	make
-	rm push_swap
+C_BIN=./checker_Mac
+if [ "$(uname)" = "Linux" ]; then
+	C_BIN=./checker
+	if [ ! -f "$C_BIN" ]; then
+		make
+		rm push_swap
+	fi
 fi
-if [ ! -f "$BIN" ]; then
-	make bonus
-fi
+make debug >/dev/null
 ./push_swap -r "$1" > tmp
-./checker_Mac $(cat .randstack) < tmp
+./$C_BIN $(cat .randstack) < tmp
 rm .randstack tmp
